fix: compilation errors - GradientDrawable imports, SceneView API, type refs
- Add missing GradientDrawable imports to 6 UI files
- Fix AvatarConfig.kt Long→Int literal defaults (.toInt())
- Fix DayNightOverlay.kt missing kotlin.math.sin import + 255→255f ambiguity
- Rewrite PetNode.kt to use correct SceneView CubeNode/SphereNode constructors
- Fix CityActivity.kt: MaterialLoader.MaterialInstance→Filament MaterialInstance,
  poleMat scope, wm.color/lampLightMaterial?.color → setParameter API
- Fix JobsActivity.kt: JobDefs→JobDef type reference
- Fix BuildingInteriorActivity.kt: add Context import, toMutableUI→toMutableMap

+++ app/src/main/java/com/intelligame/huntix/reallife/AvatarConfig.kt
@@ -16,13 +16,13 @@ import android.graphics.Shader
  */
 data class AvatarConfig(
     val gender: String = "male",         // "male" | "female"
-    val skinColor: Int = 0xFFDBA97A,     // colore pelle
-    val hairColor: Int = 0xFF3B2314,     // colore capelli
+    val skinColor: Int = 0xFFDBA97A.toInt(),     // colore pelle
+    val hairColor: Int = 0xFF3B2314.toInt(),     // colore capelli
     val hairStyle: String = "short",     // "short" | "long" | "bald" | "mohawk" | "ponytail"
-    val shirtColor: Int = 0xFFE53935,    // colore maglia
-    val pantsColor: Int = 0xFF1A237E,    // colore pantaloni
-    val shoeColor: Int = 0xFF424242,     // colore scarpe
-    val eyeColor: Int = 0xFF1B5E20       // colore occhi
+    val shirtColor: Int = 0xFFE53935.toInt(),    // colore maglia
+    val pantsColor: Int = 0xFF1A237E.toInt(),    // colore pantaloni
+    val shoeColor: Int = 0xFF424242.toInt(),     // colore scarpe
+    val eyeColor: Int = 0xFF1B5E20.toInt()       // colore occhi
 ) {
     companion object {
         private const val PREFS = "rl_avatar"
